moved Indexer from core.abc to core.containers
