+++ yaw/core/abc.py
@@ -1,9 +1,8 @@
 from __future__ import annotations
 
 from abc import ABC, abstractclassmethod, abstractmethod, abstractproperty
-from collections.abc import Iterator
 from dataclasses import asdict
-from typing import TYPE_CHECKING, Any, Callable, Generic, Type, TypeVar
+from typing import TYPE_CHECKING, Any, Type, TypeVar
 
 import h5py
 import numpy as np
@@ -12,58 +11,6 @@ from numpy.typing import NDArray
 if TYPE_CHECKING:  # pragma: no cover
     from pandas import IntervalIndex
     from yaw.core.utils import TypePathStr
-
-
-_TK = TypeVar("_TK")
-_TV = TypeVar("_TV")
-
-
-class Indexer(Generic[_TK, _TV], Iterator):
-    """Helper class to implemented a class attribute that can be used as
-    indexer and iterator for the classes stored data (e.g. indexing patches or
-    redshift bins).
-    """
-
-    def __init__(self, inst: _TV, builder: Callable[[_TV, _TK], _TV]) -> None:
-        """Construct a new indexer.
-
-        Args:
-            inst:
-                Class instance on which the indexing operations are applied.
-            builder:
-                Callable signature ``builder(inst, item) -> inst`` that
-                constructs a new class instance with the indexing specified from
-                ``item`` applied.
-
-
-        The resulting indexer supports indexing and slicing (depending on the
-        subclass implementation), as well as iteration, where instances holding
-        individual items are yielded.
-        """
-        self._inst = inst
-        self._builder = builder
-        self._iter_loc = 0
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}({self._inst.__class__.__name__})"
-
-    def __getitem__(self, item: _TK) -> _TV:
-        return self._builder(self._inst, item)
-
-    def __next__(self) -> _TV:
-        """Returns the next value and increments the iterator location index."""
-        try:
-            val = self[self._iter_loc]
-        except IndexError:
-            raise StopIteration
-        else:
-            self._iter_loc += 1
-            return val
-
-    def __iter__(self) -> Iterator[_TV]:
-        """Returns a new instance of this class to have an independent iterator
-        location index"""
-        return self.__class__(inst=self._inst, builder=self._builder)
 
 
 class PatchedQuantity(ABC):
@@ -82,7 +29,7 @@ class BinnedQuantity(ABC):
     """Base class for an object that has data organised in redshift bins."""
 
     def get_binning(self) -> IntervalIndex:
-        """Get the redshift binning of the correlation function.
+        """Get the underlying, exact redshift bin intervals.
 
         Returns:
             :obj:`pandas.IntervalIndex`
@@ -191,8 +138,7 @@ class HDFSerializable(ABC):
 
     @classmethod
     def from_file(cls: Type[_Thdf], path: TypePathStr) -> _Thdf:
-        """Create a class instance by deserialising data from the root of a HDF5
-        file.
+        """Create a class instance by deserialising data from a HDF5 file.
 
         Args:
             path (:obj:`pathlib.Path`, str):
